pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deerflow-agent-enforcer"
version = "2.0.0"
description = "Strict AI Agent Quality Framework based on DeerFlow Agentic Workflow"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-subprocess>=1.5",
    "pytest-randomly>=3.15",
]

[tool.setuptools.packages.find]
include = ["deerflow_core*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests (isolated, fast)",
    "integration: Integration tests (cross-module)",
    "slow: Tests that take > 5 seconds",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["deerflow_core"]
branch = true
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_empty = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["E501"]
